compare: Use int16 instead of int32 arrays.

+++ lib/matplotlib/testing/compare.py
@@ -307,8 +307,8 @@ def compare_images( expected, actual, tol, in_decorator=False ):
 
    # convert to signed integers, so that the images can be subtracted without
    # overflow
-   expectedImage = expectedImage.astype(np.int32)
-   actualImage = actualImage.astype(np.int32)
+   expectedImage = expectedImage.astype(np.int16)
+   actualImage = actualImage.astype(np.int16)
 
    rms = calculate_rms(expectedImage, actualImage)
